Moved all Swift extensions to 1 file
